Some Updates
Added a calendar which I have to figure out how to trigger a smiley face (svg?) to show consistency in journal entries.

Added header and footer, as a 'profile' will be needed for usage, footer for semantics.

Realizing I need to figure out how to create a user login page, then the profile display, need to sketch that out for clarity.