Sidebars
This would be my first real round dealing with positioning. My goal is to have the left panel only come out when a person clicks on 'View' to then display the full entry. This panel floatin makes the most sense, as I don't want it to get tangled in the main blog container.

The right sidebar panel is supposed to Push items to the left and be apart of the same container as the blog. It should be toggable.

I'm seeing the difference between flex and positioning, and how they do not go together with the same element selectors.

So now, I am trying to figure out how to apply what I learned on codepen for the right panel sidebar, to the left, but truthfully I think using javascript to make the right panel toggle collapse would be easier, and then i can focus using positioning on the left panel.