Changes to file name
I didnt like how the index.html page kept popping up first so I changed the name to main.html. It was only named index.html originally due to that being the first page I worked on at the start of this project. I was thinking of putting landing.html to index.html but I'll see, because it makes sense to leave the pages as what their core is.