logout works, redirection works.
the user header now shows when a user is logged in and the guest header is hidden.

the first page seen is the landing.html page (still need to work on the css to tighten it up)

just realized, i need for when a user registers, the regsiter.html goes to the main.index page

the logout button works and redirects to landing.html

i still need to use uid to connect to the speific user.

so happy, great progress!!

ive already been coding on codepen and have it within my notion notes on how the profile.html and main.html will connect and display the journal entries, i will likely add it tomorrow if i dont have time after dinner tonight.