Firebase Collections
I tried creating a new document for my collection, but every time I try registering a new user to test it, it doesnt work or I kept running into errors of variables not being defined.

I haven't had a chance to go super in depth with async and await, but I learned that they work as a pair,

async → allows the function to wait for asynchronous operations

await → actually pauses the code until the asynchronous task completes

I think I need to take some time to leaarn more on the noSQL way firebase stories information. and I think there's something involved in users that I am missing, so I will do my research and fix this issue.

Then I can move on to the main.html page to ensure when a user clicks on 'save journal', it saves and profile.html displays the previously stored data.